Print thread ID in panic message if thread name is unknown
`panic!` does not print any identifying information for threads that are
unnamed. However, in many cases, the thread ID can be determined.

This changes the panic message from something like this:

    thread '<unnamed>' panicked at src/main.rs:3:5:
    explicit panic

To something like this:

    thread '<unnamed>' (id 2) panicked at src/main.rs:3:5:
    explicit panic

There is no change in output for named threads or for threads where an
ID cannot be determined.